ubah sistem delete service charge

+++ tt_reservation_visa/models/tt_provider_visa.py
@@ -139,8 +139,13 @@ class TtProviderVisa(models.Model):
                 print(scs_obj)
 
     def delete_service_charge(self):
-        for rec in self.cost_service_charge_ids:
-            rec.unlink()
+        ledger_created = False
+        for rec in self.cost_service_charge_ids.filtered(lambda x: x.is_extra_fees == False):
+            if rec.is_ledger_created:
+                ledger_created = True
+            else:
+                rec.unlink()
+        return ledger_created
 
     def action_create_ledger(self, issued_uid, pay_method=None):
         self.env['tt.ledger'].action_create_ledger(self, issued_uid)
